Migrate image_generation to static metadata
