+++ front/lib/actions/mcp_internal_actions/servers/index.ts
@@ -24,7 +24,6 @@ import { default as driveServer } from "@app/lib/actions/mcp_internal_actions/se
 import { default as sheetsServer } from "@app/lib/actions/mcp_internal_actions/servers/google_sheets";
 import { default as httpClientServer } from "@app/lib/actions/mcp_internal_actions/servers/http_client";
 import { default as hubspotServer } from "@app/lib/actions/mcp_internal_actions/servers/hubspot";
-import { default as imageGenerationServer } from "@app/lib/actions/mcp_internal_actions/servers/image_generation";
 import { default as includeDataServer } from "@app/lib/actions/mcp_internal_actions/servers/include";
 import { default as interactiveContentServer } from "@app/lib/actions/mcp_internal_actions/servers/interactive_content";
 import { default as jiraServer } from "@app/lib/actions/mcp_internal_actions/servers/jira";
@@ -72,6 +71,7 @@ import { default as agentCopilotAgentStateServer } from "@app/lib/api/actions/se
 import { default as agentCopilotContextServer } from "@app/lib/api/actions/servers/agent_copilot_context";
 import { default as githubServer } from "@app/lib/api/actions/servers/github";
 import { default as calendarServer } from "@app/lib/api/actions/servers/google_calendar";
+import { default as imageGenerationServer } from "@app/lib/api/actions/servers/image_generation";
 import { default as snowflakeServer } from "@app/lib/api/actions/servers/snowflake";
 import type { Authenticator } from "@app/lib/auth";
 import { assertNever } from "@app/types";
