+++ front/lib/api/actions/servers/image_generation/helpers.ts
@@ -1,0 +1,261 @@
+import { GoogleGenAI } from "@google/genai";
+import { z } from "zod";
+
+import { MCPError } from "@app/lib/actions/mcp_errors";
+import type { MCPProgressNotificationType } from "@app/lib/actions/mcp_internal_actions/output_schemas";
+import { computeTokensCostForUsageInMicroUsd } from "@app/lib/api/assistant/token_pricing";
+import type { Authenticator } from "@app/lib/auth";
+import { rateLimiter } from "@app/lib/utils/rate_limiter";
+import { getStatsDClient } from "@app/lib/utils/statsd";
+import logger from "@app/logger/logger";
+import { dustManagedCredentials, Err, Ok } from "@app/types";
+import { GEMINI_2_5_FLASH_IMAGE_MODEL_ID } from "@app/types/assistant/models/google_ai_studio";
+
+export const IMAGE_GENERATION_RATE_LIMITER_KEY = "image_generation";
+export const IMAGE_GENERATION_RATE_LIMITER_TIMEFRAME_SECONDS = 60 * 60 * 24 * 7; // 1 week.
+
+export const DEFAULT_IMAGE_OUTPUT_FORMAT = "png";
+export const DEFAULT_IMAGE_MIME_TYPE = "image/png";
+
+// Token pricing is expressed as cost per million tokens (micro-USD per token)
+const MICRO_USD_PER_USD = 1_000_000;
+
+// Map tool size parameters to Gemini aspect ratios.
+export const SIZE_TO_ASPECT_RATIO: Record<string, string> = {
+  "1024x1024": "1:1",
+  "1536x1024": "3:2",
+  "1024x1536": "2:3",
+};
+
+const GeminiInlineDataPartSchema = z.object({
+  inlineData: z.object({
+    data: z.string(),
+    mimeType: z.string().optional(),
+  }),
+});
+
+export type GeminiInlineDataPart = z.infer<typeof GeminiInlineDataPartSchema>;
+
+// Type guard to validate Gemini inline data parts.
+export function isValidGeminiInlineDataPart(
+  part: unknown
+): part is GeminiInlineDataPart {
+  return GeminiInlineDataPartSchema.safeParse(part).success;
+}
+
+/**
+ * Computes cost details for Gemini image generation from usage metadata.
+ * Returns structured cost information for Langfuse observation updates.
+ */
+export function computeImageGenerationCostDetails(usageMetadata: {
+  promptTokenCount?: number;
+  candidatesTokenCount?: number;
+}): {
+  inputTokens: number;
+  outputTokens: number;
+  totalTokens: number;
+  costUsd: number;
+  costDetails: {
+    input: number;
+    output: number;
+    total: number;
+  };
+} {
+  const inputTokens = usageMetadata.promptTokenCount ?? 0;
+  const outputTokens = usageMetadata.candidatesTokenCount ?? 0;
+  const totalTokens = inputTokens + outputTokens;
+
+  const totalCostMicroUsd = computeTokensCostForUsageInMicroUsd({
+    modelId: GEMINI_2_5_FLASH_IMAGE_MODEL_ID,
+    promptTokens: inputTokens,
+    completionTokens: outputTokens,
+    cachedTokens: null,
+    cacheCreationTokens: null,
+  });
+
+  // Convert micro-USD to USD for Langfuse
+  const costUsd = totalCostMicroUsd / MICRO_USD_PER_USD;
+
+  // Proportional cost breakdown for input/output
+  const inputCostUsd =
+    totalTokens > 0 ? (costUsd * inputTokens) / totalTokens : 0;
+  const outputCostUsd =
+    totalTokens > 0 ? (costUsd * outputTokens) / totalTokens : 0;
+
+  return {
+    inputTokens,
+    outputTokens,
+    totalTokens,
+    costUsd,
+    costDetails: {
+      input: inputCostUsd,
+      output: outputCostUsd,
+      total: costUsd,
+    },
+  };
+}
+
+export async function sendImageProgressNotification(
+  sendNotification: (
+    notification: MCPProgressNotificationType
+  ) => Promise<void>,
+  progressToken: string | number | undefined,
+  label: string
+): Promise<void> {
+  if (!progressToken) {
+    return;
+  }
+
+  const notification: MCPProgressNotificationType = {
+    method: "notifications/progress",
+    params: {
+      progress: 0,
+      total: 1,
+      progressToken,
+      data: {
+        label,
+        output: {
+          type: "image",
+          mimeType: DEFAULT_IMAGE_MIME_TYPE,
+        },
+      },
+    },
+  };
+
+  await sendNotification(notification);
+}
+
+export async function checkImageGenerationRateLimit(
+  auth: Authenticator,
+  workspace: { sId: string }
+): Promise<Ok<void> | Err<MCPError>> {
+  const { limits } = auth.getNonNullablePlan();
+  const { maxImagesPerWeek } = limits.capabilities.images;
+
+  const remaining = await rateLimiter({
+    key: `${IMAGE_GENERATION_RATE_LIMITER_KEY}_${workspace.sId}`,
+    maxPerTimeframe: maxImagesPerWeek,
+    timeframeSeconds: IMAGE_GENERATION_RATE_LIMITER_TIMEFRAME_SECONDS,
+    logger,
+  });
+
+  if (remaining <= 0) {
+    const statsDClient = getStatsDClient();
+    statsDClient.increment("tools.image_generation.rate_limit_hit", 1, [
+      "provider:gemini",
+    ]);
+
+    return new Err(
+      new MCPError(
+        `Rate limit of ${maxImagesPerWeek} requests per week exceeded. Contact your ` +
+          "administrator to increase the limit.",
+        {
+          tracked: false,
+        }
+      )
+    );
+  }
+
+  return new Ok(undefined);
+}
+
+export function validateGeminiImageResponse(
+  response: {
+    candidates?: Array<{
+      content?: {
+        parts?: unknown[];
+      };
+    }>;
+    promptFeedback?: {
+      blockReason?: string;
+      safetyRatings?: unknown;
+    };
+  },
+  operationType: "generation" | "editing",
+  promptText: string
+): Ok<GeminiInlineDataPart[]> | Err<MCPError> {
+  // Check for empty candidates.
+  if (!response.candidates || response.candidates.length === 0) {
+    if (response.promptFeedback?.blockReason) {
+      logger.error(
+        {
+          blockReason: response.promptFeedback.blockReason,
+          safetyRatings: response.promptFeedback.safetyRatings,
+          prompt: promptText,
+        },
+        `Gemini image ${operationType}: Prompt blocked by safety filters`
+      );
+      return new Err(
+        new MCPError(
+          `Image ${operationType} blocked by safety filters: ${response.promptFeedback.blockReason}`
+        )
+      );
+    }
+    return new Err(new MCPError("No image generated."));
+  }
+
+  // Validate content structure.
+  const content = response.candidates[0].content;
+  if (!content || !content.parts) {
+    return new Err(new MCPError("No image data in response"));
+  }
+
+  // Extract valid image parts.
+  const imageParts = content.parts.filter(isValidGeminiInlineDataPart);
+  if (imageParts.length === 0) {
+    return new Err(new MCPError("No image data in response."));
+  }
+
+  return new Ok(imageParts);
+}
+
+export function trackGeminiTokenUsage(
+  response: {
+    usageMetadata?: {
+      promptTokenCount?: number;
+      candidatesTokenCount?: number;
+    };
+  },
+  statsDClient: ReturnType<typeof getStatsDClient>
+): void {
+  statsDClient.increment(
+    "tools.image_generation.usage.input_tokens",
+    response.usageMetadata?.promptTokenCount ?? 0,
+    ["provider:gemini"]
+  );
+  statsDClient.increment(
+    "tools.image_generation.usage.output_tokens",
+    response.usageMetadata?.candidatesTokenCount ?? 0,
+    ["provider:gemini"]
+  );
+}
+
+export function formatImageResponse(
+  imageParts: GeminiInlineDataPart[],
+  fileName: string
+): Ok<
+  Array<{
+    type: "image";
+    mimeType: string;
+    data: string;
+    name: string;
+  }>
+> {
+  const outputFileName = `${fileName}.${DEFAULT_IMAGE_OUTPUT_FORMAT}`;
+
+  return new Ok(
+    imageParts.map((part) => ({
+      type: "image" as const,
+      mimeType: part.inlineData.mimeType ?? DEFAULT_IMAGE_MIME_TYPE,
+      data: part.inlineData.data,
+      name: outputFileName,
+    }))
+  );
+}
+
+export function createGeminiClient(): GoogleGenAI {
+  const credentials = dustManagedCredentials();
+  return new GoogleGenAI({
+    apiKey: credentials.GOOGLE_AI_STUDIO_API_KEY,
+  });
+}
